Question answer
The user clicks enter to move to the next screen then the first question is drawn with 2 answer options, when the user presses the left mouse button the screen changes. Next i have to ensure that the screen will change to a new question, noot just a blank background. I also must ensure that when the correct option is clicked (not just anywhere on the screen) the game changes. I will then include elements such as levels to the main game, so that the questions get harder the more the user gets right etc.