Cleaned up my files//Code now works
I implemented the layer template code into my text input code but it still came up with the error that FtLayer could not be found. I tabbed out all my own code and ran it again, this produced the same error so I checked the folder and I had copied some of the module files in there which was confusing the program. I removed them from the folder and now I have solved that error.